// types.ts
export interface FormData {
  name: string;
  phoneNumber: string;
  eventType: string;
  eventDate: string;
}

export interface ServiceItem {
  id: string;
  title: string;
  description: string;
  image: string;
}

export interface WhyUsItem {
  icon: string;
  title: string;
  description: string;
}